# perf: add obliterator to benchmarks