--- a/package.json
+++ b/package.json
@@ -85,6 +85,7 @@
     "lodash": "^4.17.11",
     "mocha": "^6.1.4",
     "nyc": "^14.1.1",
+    "obliterator": "github:forivall/obliterator#patch-1",
     "prettier": "^1.18.1",
     "rxjs": "^6.5.2",
     "semantic-release": "^15.13.12",