I was curious about the performance compared to [obliterator](https://github.com/Yomguithereal/obliterator) (using my fork to fix the type definitions), so i added it to the benchmark. You're still the fastest lib.

```
~pubrepos/iterare master*
» node ./lib/benchmarks/filter_take_set.js
Loop x 2,102,039 ops/sec ±1.37% (86 runs sampled)
iterare x 603,265 ops/sec ±2.23% (82 runs sampled)
Array method chain x 43.79 ops/sec ±1.17% (57 runs sampled)
Lodash x 642 ops/sec ±1.19% (85 runs sampled)
RxJS x 110,212 ops/sec ±4.47% (88 runs sampled)
IxJS (ES5) x 359,412 ops/sec ±1.11% (89 runs sampled)
IxJS (ES6) x 325,086 ops/sec ±2.87% (82 runs sampled)
obliterator x 416,279 ops/sec ±1.95% (87 runs sampled)
Fastest is Loop

~pubrepos/iterare master* 49s
» node ./lib/benchmarks/map_filter_set.js
Loop x 408 ops/sec ±1.96% (83 runs sampled)
iterare x 382 ops/sec ±1.06% (83 runs sampled)
Array method chain x 340 ops/sec ±5.81% (84 runs sampled)
Lodash x 363 ops/sec ±1.36% (85 runs sampled)
RxJS x 342 ops/sec ±1.36% (84 runs sampled)
IxJS (ES5) x 201 ops/sec ±2.41% (80 runs sampled)
IxJS (ES6) x 212 ops/sec ±1.09% (81 runs sampled)
obliterator x 258 ops/sec ±1.29% (79 runs sampled)
Fastest is Loop
```